+++ .changeset/tasty-beans-clap.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-backend-module-bitbucket': patch
+---
+
+split BitbucketClient into BitbucketCloudClient, BitbucketServerClient
