+++ plugins/catalog-backend-module-bitbucket/src/lib/index.ts
@@ -16,6 +16,8 @@
 
 export { defaultRepositoryParser } from './BitbucketRepositoryParser';
 export type { BitbucketRepositoryParser } from './BitbucketRepositoryParser';
-export { BitbucketClient, paginated, paginated20 } from './client';
-export type { PagedResponse, PagedResponse20 } from './client';
+export { BitbucketCloudClient, paginated20 } from './BitbucketCloudClient';
+export { BitbucketServerClient, paginated } from './BitbucketServerClient';
+export type { PagedResponse20 } from './BitbucketCloudClient';
+export type { PagedResponse } from './BitbucketServerClient';
 export type { BitbucketRepository, BitbucketRepository20 } from './types';
